Rename components and pages files
