+++ client/src/components/AnimatedCounter.component.js
@@ -12,11 +12,17 @@ export default function AnimatedCounter({ from, to }) {
   const rounded = useTransform(count, (latest) => Math.round(latest));
   const ref = useRef(null);
   const inView = useInView(ref);
+
   // while in view, animate the count
   useEffect(() => {
     if (inView) {
       animate(count, to, { duration: 3 });
     }
   }, [count, inView, to]);
-  return <motion.span ref={ref}>{rounded}</motion.span>;
+
+  return (
+    <>
+      <motion.span ref={ref}>{rounded}</motion.span>;
+    </>
+  );
 }
