+++ client/src/components/Banner.component.js
