+++ client/src/components/Footer.component.js
