+++ client/src/components/Stats.component.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef } from "react";
 import { Background, Parallax } from "react-parallax";
 import LseBanner from "../assets/img/LseBanner.webp";
-import AnimatedCounter from "./AnimatedCounter";
+import AnimatedCounter from "./AnimatedCounter.component";
 import { motion, useInView, useAnimation } from "framer-motion";
 
 export default function Stats() {
