--- client/src/views/Home.js
@@ -1,18 +1,0 @@
-import React from "react";
-
-//Components
-import Header from "../components/Header";
-import Banner from "../components/Banner";
-import Stats from "../components/Stats";
-import Footer from "../components/Footer";
-
-export default function Home() {
-  return (
-    <>
-      <Header />
-      <Banner />
-      <Stats />
-      <Footer />
-    </>
-  );
-}
